Done: test difference of keeping state between sessions
